split replay plugin and module
